Introduce the Recommends dependency
Pulp has been ignoring the Recommends weak dependency when processing e.g
recursive unit copies.  This patch updates the RPM model to track a
recommends unit attribute and the primary.xml parser to populate this
attribute.  The libsolv dependency solver is used to process this
attribute when calculating unit dependencies.

Fixes: #3847
https://pulp.plan.io/issues/3847